feat: make GetForm query a composable
