+++ frontend/pages/studies/[studyId]/[submissionId]/[slug].vue
@@ -1,131 +1,11 @@
 <script lang="ts" setup>
-import { useQuery } from "@vue/apollo-composable";
-import { graphql } from "~/generated/gql";
-import type { GetFormQuery } from "~/generated/gql/graphql";
+import { useFormQuery } from "~/composables/useFormQuery";
 import FormWrapper from "~/components/form/FormWrapper.vue";
 import { useStepSlug, useSubmissionId } from "~/composables/useRouteParams";
 
-const GET_FORM = graphql(`
-    query GetForm($submissionId: ID!) {
-        form(submissionId: $submissionId, mrPermission: "Edit") {
-            formId
-            nameEn
-            nameNl
-            submissionId
-            steps {
-                stepId
-                slug
-                repeatIndex
-                nameEn
-                nameNl
-                descriptionEn
-                descriptionNl
-                isOverview
-                ...FormInfoFragment
-                questions {
-                    questionId
-                    repeatIndex
-                    answer
-                    responseId
-                    textEn
-                    textNl
-                    descriptionEn
-                    descriptionNl
-                    required
-                    hasConditions
-                    ... on SelectQuestionType {
-                        multiple
-                        options {
-                            id
-                            labelNl
-                            labelEn
-                            defaultSelected
-                        }
-                    }
-                    ... on NumberQuestionType {
-                        positiveOnly
-                    }
-                    ... on TrueFalseQuestionType {
-                        defaultValue
-                    }
-                    ... on FileUploadQuestionType {
-                        sizeLimit
-                    }
-                    ... on TextQuestionType {
-                        placeholderNl
-                        placeholderEn
-                        lines
-                        isEmail
-                    }
-                    ... on DateQuestionType {
-                        futureOnly
-                    }
-                }
-                substeps {
-                    stepId
-                    slug
-                    repeatIndex
-                    nameEn
-                    nameNl
-                    descriptionEn
-                    descriptionNl
-                    isOverview
-                    ...FormInfoFragment
-                    questions {
-                        questionId
-                        repeatIndex
-                        answer
-                        responseId
-                        textEn
-                        textNl
-                        descriptionEn
-                        descriptionNl
-                        required
-                        hasConditions
-                        ... on SelectQuestionType {
-                            multiple
-                            options {
-                                id
-                                labelNl
-                                labelEn
-                                defaultSelected
-                            }
-                        }
-                        ... on NumberQuestionType {
-                            positiveOnly
-                        }
-                        ... on TrueFalseQuestionType {
-                            defaultValue
-                        }
-                        ... on FileUploadQuestionType {
-                            sizeLimit
-                        }
-                        ... on TextQuestionType {
-                            placeholderNl
-                            placeholderEn
-                            lines
-                            isEmail
-                        }
-                        ... on DateQuestionType {
-                            futureOnly
-                        }
-                    }
-                }
-            }
-        }
-    }
-`);
-
 const submissionId = useSubmissionId();
 const slug = useStepSlug();
-
-const { result: formResult } = useQuery<GetFormQuery>(
-    GET_FORM,
-    () => ({ submissionId: submissionId.value }),
-    () => ({ enabled: !!submissionId.value }),
-);
-
-const form = computed(() => formResult.value?.form ?? null);
+const form = useFormQuery(submissionId, "Edit");
 </script>
 
 <template>
